refactor(client): Move all config client tests to same file
